Crate recon_mcts_test_nim

A recon_mcts API demonstration of a simple Nim variant. See code.

The game involves two players, a cumulative value (initially set to 500 in this example), and a maximum move value (10 in this example). The players take turns to reach 0 by subtracting a move value between 1 and 10 from the cumulative value. The first player to reach 0 wins.

This game presents an interesting demonstration for tree search because it is intractable for a non-recombining tree. In the example given above, the search space for a non-recombining tree is greater than 1050.

Implementation Note

Note that this implementation is designed to be a demonstration of the API and therefore less succinct than it could be. For example, both static and dynamic dispatch version of the game are implemented via GameDynamics and DynGD, respectively.